backend/requirements.txt
langchain
langchain-huggingface
langchain-text-splitters
langchain-unstructured
langchain-openai
faiss-cpu
unstructured
Markdown
sentence-transformers
lazy_loader
rank_bm25
openai
fastapi
uvicorn
python-dotenv
pydantic
jieba>=0.42.1
sqlalchemy
aiosqlite
